Animation editor enhancements
Bone positioning now takes into account its parent's absolute rotation,
and compensates it. That means bones with rotated parents follow exactly
the mouse when dragged, instead of going anywhere except where they should.

Repaired selecting bones with the mouse, and made that the default
(can be switched to keyboard with M key).

The timeline grid size and timestep unit size are now variable,
and can be changed with the U, I, O, P keys or the added UI buttons.

Bone borders and joint points can be displayed with B key.

Removed the ignorebone0 button and related functionality.

Minor cosmetical things.

+++ BBGE/Vector.h
@@ -402,7 +402,7 @@ public:
 	 }
 #endif
 	 void rotate2DRad(float rad);
-	 void rotate2D360(int angle);
+	 void rotate2D360(float angle);
 };
